chore(ui-components): refacfor useSupport
