+++ packages/ui-components/src/layouts/Forms/AutoForm.stories.tsx
@@ -3,14 +3,14 @@ import React from 'react';
 
 import { ThemeProvider } from '../..';
 import AutoForm from './AutoForm';
-import { autoFormSchema } from './autoFormSchema';
+import { autoFormSchema } from './validations';
 
 export default {
   title: 'UI Layouts/Form/AutoForm',
   component: AutoForm,
 } as ComponentMeta<typeof AutoForm>;
 
-const Template: ComponentStory<typeof AutoForm> = args => (
+const Template: ComponentStory<typeof AutoForm> = (args) => (
   <ThemeProvider>
     <AutoForm {...args} />
   </ThemeProvider>
