+++ packages/ui-components/src/layouts/Forms/AutoForm.test.tsx
@@ -1,8 +1,8 @@
-import AutoForm from './AutoForm';
 import '@testing-library/jest-dom';
 import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
-import { autoFormSchema } from './autoFormSchema';
+import AutoForm from './AutoForm';
+import { autoFormSchema } from './validations';
 
 describe('Auto form component', () => {
   beforeEach(() => {
